requirements.txt
# Core
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.26.2

# Alpaca API
alpaca-py==0.13.3
websocket-client==1.7.0

# APIs
requests==2.31.0
aiohttp==3.9.1

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Telegram
python-telegram-bot==20.7

# Technical Analysis
ta==0.11.0
yfinance==0.2.33

# AI/ML
openai==1.6.1
anthropic==0.8.1

# Utils
pytz==2023.3
schedule==1.2.0
colorlog==6.8.0

# Development
pytest==7.4.3
black==23.12.1
flake8==7.0.0
ipython==8.19.0